RectangleTexture and createRenderTexture util
