+++ src/renderers/webgl/textures/ArrowTexture.ts
@@ -1,7 +1,8 @@
-import { RenderTexture, Graphics, Matrix, MSAA_QUALITY, Renderer as PixiRenderer } from 'pixi.js'
+import { RenderTexture, Graphics, Matrix } from 'pixi.js'
+import { createRenderTexture } from '../../../utils/webgl'
 import { MIN_TEXTURE_ZOOM } from '../../../utils/constants'
-import { Texture } from '../../../types'
 import { Renderer } from '..'
+import { Texture } from '../../../types'
 
 export default class ArrowTexture implements Texture {
   private texture: RenderTexture
@@ -12,23 +13,7 @@ export default class ArrowTexture implements Texture {
       .lineTo(this.height * this.scaleFactor, this.width * this.scaleFactor * 0.5)
       .lineTo(this.height * this.scaleFactor, -this.width * this.scaleFactor * 0.5)
 
-    this.texture = RenderTexture.create({
-      width: graphic.width,
-      height: graphic.height,
-      multisample: MSAA_QUALITY.HIGH,
-      resolution: 2
-    })
-
-    renderer.app.renderer.render(graphic, {
-      renderTexture: this.texture,
-      transform: new Matrix(1, 0, 0, 1, 0, graphic.height / 2)
-    })
-
-    if (renderer.app.renderer instanceof PixiRenderer) {
-      renderer.app.renderer.framebuffer.blit()
-    }
-
-    graphic.destroy(true)
+    this.texture = createRenderTexture(renderer.app, graphic, new Matrix(1, 0, 0, 1, 0, graphic.height / 2))
   }
 
   get() {
